### 2025-04-23 #8095 @danielskatz
👋 @SKopecz - thanks for your submission.

While I look for an editor, you could work on the possibly missing DOI that editorialbot suggests, but note that it may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.
