### 2025-04-23 65f6040
Added missing DOIs
