+++ paper/paper.bib
@@ -290,7 +290,8 @@
   number={3},
   pages={237--245},
   year={1978},
-  publisher={EDP Sciences}
+  publisher={EDP Sciences},
+  doi = {10.1051/m2an/1978120302371}
 }
 
 @book{gottlieb2011strong,
@@ -299,7 +300,8 @@
   author={Gottlieb, Sigal and Ketcheson, David I and Shu, Chi-Wang},
   year={2011},
   publisher={World Scientific},
-  address={Singapore}
+  address={Singapore},
+  doi = {10.1142/7498}
 }
 
 @article {Bruggeman2007,
