### 2025-04-24 #8095 @junghans
@kellertuer, I asked you because of your experience in Julia packages hence I am very sure you could give some valuable feedback in this review.
