### 2025-04-25 #8095 @kellertuer
Hm, give me a few days to think about it (or to finish off all administrative things for our semester here). I would consider myself only mildly experienced in Julia package review.
