### 2025-04-28 #8130 @editorialbot
[REVIEW]: PositiveIntegrators.jl: A Julia library of positivity-preserving time integration methods
**Submitting author:** @SKopecz (<a href="http://orcid.org/0000-0003-0593-450X">Stefan Kopecz</a>)
**Repository:** https://github.com/NumericalMathematics/PositiveIntegrators.jl
**Branch with paper.md** (empty if default branch): paper-2025-joss
**Version:** v0.2.11
