### 2025-04-29 f0a38fd
no runtime-dispatch for MPDeC + StaticArrays and first benchmarks npzd model
