+++ src/mpdec.jl
@@ -1,3 +1,21 @@
+# Helper functions
+
+function cmatrix(uprev, ::Val{M})  where {M}
+    return Array{eltype(uprev)}(undef, length(uprev), M+1)
+end
+function cmatrix(::SVector{N, T}, ::Val{M}) where {N, T, M}
+    return MMatrix{N, M+1, T}(undef)
+end
+
+# Helper function which supports creation of a square matrix from SVector
+# See https://discourse.julialang.org/t/generic-square-matrix-constructor-from-a-vector/62306
+function sqmatr(vec::AbstractVector{T}, F::DataType = T) where {T}
+    axis, = axes(vec)
+    return similar(vec, F, (axis, axis))
+end
+
+#####################################################################
+
 """
     MPDeC(K; [nodes = :gausslobatto, linsolve = ..., small_constant = ...])
 
@@ -217,12 +235,13 @@ function get_constant_parameters(alg::MPDeC, type)
     return nodes, theta
 end
 
-struct MPDeCConstantCache{NType, T, T2} <: OrdinaryDiffEqConstantCache
+struct MPDeCConstantCache{NType, T, T2,TM} <: OrdinaryDiffEqConstantCache
     K::Int
     M::Int
     nodes::NType
     theta::T2
     small_constant::T
+    ValM::TM
 end
 
 # Out-of-place
@@ -236,7 +255,7 @@ function alg_cache(alg::MPDeC, u, rate_prototype, ::Type{uEltypeNoUnits},
 
     nodes, theta = get_constant_parameters(alg, uEltypeNoUnits)
     MPDeCConstantCache(alg.K, alg.M, nodes, theta,
-                       alg.small_constant_function(uEltypeNoUnits))
+                       alg.small_constant_function(uEltypeNoUnits),Val(alg.M))
 end
 
 function initialize!(integrator, cache::MPDeCConstantCache)
@@ -245,7 +264,7 @@ end
 # out-of-place
 function build_mpdec_matrix_and_rhs_oop(uprev, m, f, C, p, t, dt, nodes, theta,
                                         small_constant)
-    N = length(uprev)
+    #N = length(uprev)
     if f isa PDSFunction
         # Additional destruction terms
         Mmat, rhs = _build_mpdec_matrix_and_rhs_oop(uprev, m, f.p, C, p, t, dt, nodes,
@@ -258,11 +277,7 @@ function build_mpdec_matrix_and_rhs_oop(uprev, m, f, C, p, t, dt, nodes, theta,
                                                     small_constant)
     end
 
-    if uprev isa StaticArray
-        return SMatrix{N, N}(Mmat), SVector{N}(rhs)
-    else
-        return Mmat, rhs
-    end
+    return Mmat, rhs
 end
 
 # out-of-place for dense arrays
@@ -270,12 +285,8 @@ end
                                                  small_constant, dest = nothing)
     N, M_plus_1 = size(C)
 
-    # Create linear system matrix and rhs
-    if uprev isa StaticArray
-        Mmat = MMatrix{N, N}(zeros(eltype(uprev), N, N))
-    else
-        Mmat = zeros(eltype(uprev), N, N)
-    end
+    Mmat = sqmatr(uprev)
+    fill!(Mmat, zero(eltype(Mmat))) 
     rhs = similar(uprev)
 
     # Initialize
@@ -297,6 +308,11 @@ end
             d = nothing
         end
         _build_mpdec_matrix_and_rhs!(Mmat, rhs, P, dt_th, σ, d)
+    end
+
+    if uprev isa StaticArray
+        Mmat = SMatrix(Mmat)
[…]
