+++ docs/src/convergence.md
@@ -14,7 +14,7 @@ u_2' & = \sin(2\pi t)^2 u_1 - \cos(\pi t)^2 u_2, & u_2(0)&=0.1,
 ```
 for ``0≤ t≤ 1``.
 The PDS is conservative since the sum of the right-hand side terms equals zero. 
-An implementation of the problem is given next.
+An implementation of this problem is given next.
 
 
 ```@example eoc
@@ -29,17 +29,42 @@ nothing # hide
 
 To use `analyticless_test_convergence` from [DiffEqDevTools.jl](https://github.com/SciML/DiffEqDevTools.jl), we need to pick a solver to compute the reference solution and specify tolerances.
 Since the problem is not stiff, we use the high-order explicit solver `Vern9()` from [OrdinaryDiffEqVerner.jl](https://docs.sciml.ai/OrdinaryDiffEq/stable/).
-Moreover, we choose time step sizes to investigate the convergence behavior. 
-
 ```@example eoc
 using OrdinaryDiffEqVerner
 using DiffEqDevTools: analyticless_test_convergence
 
 # solver and tolerances to compute reference solution
 test_setup = Dict(:alg => Vern9(), :reltol => 1e-14, :abstol => 1e-14)
+nothing # hide
+```
 
-# choose step sizes
-dts = 0.5 .^ (5:10)
+To keep the code short, we also define an auxiliary function that outputs a convergence table. The table lists the errors obtained with the respective time step size ``Δ t`` as well as the estimated order of convergence in parentheses.
+
+```@example eoc
+using Printf: @sprintf
+using PrettyTables: pretty_table
+
+# auxiliary function
+function convergence_table(dts, prob, algs, labels, test_setup)
+
+# compute errors and estimated convergence orders
+err_eoc = []
+for i in eachindex(algs)
+     sim = analyticless_test_convergence(dts, prob, algs[i], test_setup)
+
+     err = sim.errors[:l∞]
+     eoc = [NaN; -log2.(err[2:end] ./ err[1:(end - 1)])]
+
+     push!(err_eoc, tuple.(err, eoc))
+end
+
+# gather data for table
+data = hcat(dts, reduce(hcat,err_eoc))
+
+# print table
+formatter = (v, i, j) ->  (j>1) ? (@sprintf "%5.2e (%4.2f) " v[1] v[2]) : (@sprintf "%5.2e " v)
+pretty_table(data, formatters = formatter, header = ["Δt"; labels]) 
+end
 
 nothing # hide
 ```
@@ -49,35 +74,14 @@ nothing # hide
 First, we test several second-order MPRK schemes.
 
 ```@example eoc
+# choose step sizes
+dts = 0.5 .^ (5:10)
+
 # select schemes
-algs2 = [MPRK22(0.5); MPRK22(2.0 / 3.0); MPRK22(1.0); SSPMPRK22(0.5, 1.0)]
-labels2 = ["MPRK22(0.5)"; "MPRK22(2.0/3.0)"; "MPRK22(1.0)"; "SSPMPRK22(0.5, 1.0)"]
+algs2 = [MPRK22(0.5); MPRK22(2.0 / 3.0); MPRK22(1.0); SSPMPRK22(0.5, 1.0); MPDeC(2)]
+labels2 = ["MPRK22(0.5)"; "MPRK22(2.0/3.0)"; "MPRK22(1.0)"; "SSPMPRK22(0.5, 1.0)"; "MPDeC(2)"]
 
-# compute errors and experimental order of convergence
-err_eoc = []
-for i in eachindex(algs2)
-     sim = analyticless_test_convergence(dts, prob, algs2[i], test_setup)
-
[…]
