### 2025-05-09 1ca4abd
MPDeC in NPZD benchmark
