### 2025-05-09 584c39f
MPDeC in NPZD benchmark
