### 2025-05-12 5601cc2
increased MPDeC default small_constant to run stratospheric reaction benchmark with constant time steps
