### 2025-05-13 f9f9370
mpdec in stratospheric reaction benchmark
