+++ docs/src/stratospheric_reaction_benchmark.md
@@ -146,7 +146,7 @@ nothing # hide
 
 We also note that MPRK schemes with stricter tolerances, quickly require more than a million time steps, which makes these schemes inefficient in such situations.
 
-First we compare different MPRK schemes. In addition to the default version we also use the schemes with `small_constant = 1e-6`.
+First we compare different low-order MPRK schemes. In addition to the default version we also use the schemes with `small_constant = 1e-6`.
 
 ```@example stratreac
 # choose methods to compare
@@ -161,20 +161,43 @@ labels = ["MPRK22(1.0)"; "MPRK22(1.0, sc=1e-6)"; "SSPMPRK22(0.5,1.0)"; "SSPMPRK2
 wp = work_precision_adaptive(prob, algs, labels, abstols, reltols, alg_ref; compute_error)
 
 # plot work-precision diagram
-plot(wp, labels; title = "Stratospheric reaction benchmark", legend = :bottomleft,
+plot(wp, labels; title = "Stratospheric reaction benchmark", legend = :outerright,
      color = permutedims([repeat([1],2)..., repeat([2],2)..., repeat([3],4)..., repeat([4],4)...]),
-     xlims = (10^-7, 10^0), xticks = 10.0 .^ (-8:1:0),
-     ylims = (10^-5, 10^1), yticks = 10.0 .^ (-5:1:1), minorticks = 10)
+     xlims = (10^-6, 10^0), xticks = 10.0 .^ (-6:1:0),
+     ylims = (10^-4, 10^1), yticks = 10.0 .^ (-4:1:1), minorticks = 10)
 ```
 
-We see that using `small_constant = 1e-6` clearly improves the performance of most methods.
-For comparisons with other second- and third-order schemes from [OrdinaryDiffEq.jl](https://docs.sciml.ai/OrdinaryDiffEq/stable/) we choose the second-order scheme `MPRK22(1.0, small_constant = 1e-6)` and the third-order scheme `MPRK43I(0.5, 0.75)`.
+We see that using `small_constant = 1e-6` clearly improves the performance of some methods. 
+Next, we include the `MPDeC` methods in the comparison and use `MPRK22(1.0, small_constant = 1e-6)` and `MPRK43I(1.0, 0.5)` as a reference. 
+
+```@example stratreac
+# choose methods to compare
+algs = [MPRK22(1.0, small_constant = 1e-6); MPRK43I(1.0, 0.5); 
+        MPDeC(2); MPDeC(3); MPDeC(4); MPDeC(5); MPDeC(6); MPDeC(7); MPDeC(8); MPDeC(9); MPDeC(10);
+        MPDeC(2, small_constant = 1e-6); MPDeC(3, small_constant = 1e-6); MPDeC(4, small_constant = 1e-6); MPDeC(5, small_constant = 1e-6); MPDeC(6, small_constant = 1e-6); 
+        MPDeC(7, small_constant = 1e-6); MPDeC(8, small_constant = 1e-6); MPDeC(9, small_constant = 1e-6); MPDeC(10, small_constant = 1e-6)]
+labels = ["MPRK22(1.0, sc=1e-6)"; "MPRK43I(1.0,0.5)"; 
+          "MPDeC(2)"; "MPDeC(3)"; "MPDeC(4)"; "MPDeC(5)"; "MPDeC(6)"; "MPDeC(7)"; "MPDeC(8)"; "MPDeC(9)"; "MPDeC(10)";
[…]
