+++ docs/src/convergence.md
@@ -1,7 +1,7 @@
 # [Experimental convergence order of MPRK schemes](@id convergence_mprk)
 
-In this tutorial, we check that all implemented MPRK schemes in principle show the expected order of convergence.
-We also address the issue that some methods suffer from order reduction when the solution is close to zero.
+In this tutorial, we check that all implemented MPRK schemes can achieve their expected order of convergence.
+We also address the issue that some methods suffer from order reduction when the solution gets too close to zero.
 
 ## Conservative production-destruction systems
 
@@ -39,7 +39,7 @@ test_setup = Dict(:alg => Vern9(), :reltol => 1e-14, :abstol => 1e-14)
 nothing # hide
 ```
 
-To keep the code short, we also define an auxiliary function that outputs a convergence table. The table lists the errors obtained with the respective time step size ``Δ t`` as well as the estimated order of convergence in parentheses.
+To keep the code short, we also define an auxiliary function that outputs a convergence table, which lists the errors obtained with the respective time step size ``Δ t`` as well as the estimated order of convergence in parentheses.
 
 ```@example eoc
 using Printf: @sprintf
@@ -70,38 +70,30 @@ end
 nothing # hide
 ```
 
-### Second-order MPRK schemes
+### Second-order and third-order MPRK schemes
 
-First, we test several second-order MPRK schemes.
+First, we test several second-order and third-order MPRK schemes.
 
 ```@example eoc
 # choose step sizes
 dts = 0.5 .^ (5:10)
 
-# select schemes
+# select 2nd order schemes
 algs2 = [MPRK22(0.5); MPRK22(2.0 / 3.0); MPRK22(1.0); SSPMPRK22(0.5, 1.0); MPDeC(2)]
 labels2 = ["MPRK22(0.5)"; "MPRK22(2.0/3.0)"; "MPRK22(1.0)"; "SSPMPRK22(0.5, 1.0)"; "MPDeC(2)"]
 
-convergence_table(dts, prob, algs2, labels2, test_setup)
-```
-
-The table shows that all schemes converge as expected.
-
-### Third-order MPRK schemes
-
-In this section, we consider third-order MPRK schemes.
-
-```@example eoc
 # select 3rd order schemes
 algs3 = [MPRK43I(1.0, 0.5); MPRK43I(0.5, 0.75); MPRK43II(0.5); MPRK43II(2.0 / 3.0); 
          SSPMPRK43(); MPDeC(3)]
 labels3 = ["MPRK43I(1.0,0.5)"; "MPRK43I(0.5, 0.75)"; "MPRK43II(0.5)"; "MPRK43II(2.0/3.0)";
           "SSPMPRK43()"; "MPDeC(3)"]
 
+convergence_table(dts, prob, algs2, labels2, test_setup)
+
 convergence_table(dts, prob, algs3, labels3, test_setup)
 ```
 
-As above, the table shows that all schemes converge as expected.
+The table shows that all schemes converge as expected.
 
 ### Higher-order MPRK schemes
 
@@ -158,15 +150,11 @@ nothing # hide
 
 The following tables demonstrate that the chosen MPRK schemes converge as expected also for this non-conservative PDS.
 
-### Second-order MPRK schemes
+### Second-order and third-order MPRK schemes
 
 ```@example eoc
-convergence_table(dts, prob, algs2, labels2, test_setup)             
-```
+convergence_table(dts, prob, algs2, labels2, test_setup)    
 
-### Third-order MPRK schemes
-
-```@example eoc
[…]
