### 2025-05-16 b5b0a5f
Apply suggestions from code review
Co-authored-by: Hendrik Ranocha <ranocha@users.noreply.github.com>
