### 2025-05-16 96f744f
compat entry for DoubleFloats
