+++ docs/Project.toml
@@ -26,6 +26,7 @@ DiffEqBase = "6.160"
 DiffEqCallbacks = "4"
 DiffEqDevTools = "2.45.1"
 Documenter = "1"
+DoubleFloats = "1.4"
 InteractiveUtils = "1"
 LinearAlgebra = "1"
 LinearSolve = "3.7.1"
