### 2025-05-16 850302e
removed function sqmtr
