### 2025-05-16 b599825
Use `MPDeC` in benchmarks (#160)
* no runtime-dispatch for MPDeC + StaticArrays and first benchmarks npzd model

* order reduction in convergence benchmark

* MPDeC in NPZD benchmark

* MPDeC in NPZD benchmark

* bugfix

* mpded in npzd and robertson benchmark

* increased MPDeC default small_constant to run stratospheric reaction benchmark with constant time steps

* bugfix

* mpdec in stratospheric reaction benchmark

* minor changes

* Apply suggestions from code review

Co-authored-by: Hendrik Ranocha <ranocha@users.noreply.github.com>

* compat entry for DoubleFloats

* removed function s
[…]
