+++ docs/src/robertson_benchmark.md
@@ -105,25 +105,28 @@ We start with a comparison of different adaptive MPRK schemes.
 ```@example ROBER
 # choose methods to compare
 algs = [MPRK22(0.5); MPRK22(2.0 / 3.0); MPRK22(1.0); MPRK43I(1.0, 0.5); MPRK43I(0.5, 0.75);
-        MPRK43II(0.5); MPRK43II(2.0 / 3.0)]
+        MPRK43II(0.5); MPRK43II(2.0 / 3.0);
+        MPDeC(2); MPDeC(3); MPDeC(4); MPDeC(5); MPDeC(6); MPDeC(7); MPDeC(8); MPDeC(9); MPDeC(10)]
 labels = ["MPRK22(0.5)"; "MPRK22(2/3)"; "MPRK22(1.0)"; "MPRK43I(1.0,0.5)"; "MPRK43I(0.5,0.75)";
-         "MPRK43II(0.5)"; "MPRK43II(2.0/3.0)"]
+         "MPRK43II(0.5)"; "MPRK43II(2.0/3.0)";
+         "MPDeC(2)"; "MPDeC(3)"; "MPDeC(4)"; "MPDeC(5)"; "MPDeC(6)"; "MPDeC(7)"; "MPDeC(8)"; "MPDeC(9)"; "MPDeC(10)"]
 
 # compute work-precision data
 wp = work_precision_adaptive(prob, algs, labels, abstols, reltols, alg_ref;
                             adaptive_ref = true, compute_error)
 
 # plot work-precision diagram
-plot(wp, labels; title = "Robertson benchmark", legend = :topright,
-     color = permutedims([repeat([1], 3)..., repeat([3], 2)..., repeat([4], 2)...]),
+plot(wp, labels; title = "Robertson benchmark", legend = :outerright,
+     color = permutedims([repeat([1], 3)..., repeat([3], 2)..., repeat([4], 2)..., repeat([5], 9)...]),
      xlims = (10^-10, 10^0), xticks = 10.0 .^ (-10:1:0),
      ylims = (10^-5, 10^0), yticks = 10.0 .^ (-5:1:0), minorticks = 10)
 ```
 
 We see that the second- and third-order schemes perform very similar, except for `MPRK22(0.5)`.
 This superior performance of `MPRK22(0.5)` cannot be seen in other benchmarks; it is, therefore, an exception here.
+We also see a benefit in using higher-order `MPDeC` methods.
 
-The scheme `SSPMPRK22(0.5, 1.0)` has not been considered above, since it generates oscillatory solutions.
+The scheme `SSPMPRK22(0.5, 1.0)` has not been considered above, since it generates oscillatory solutions that lead to large errors.
 
 ```@example ROBER
 sol1 = solve(prob, SSPMPRK22(0.5, 1.0), abstol=1e-5, reltol = 1e-4);
@@ -132,25 +135,27 @@ sol1 = solve(prob, SSPMPRK22(0.5, 1.0), abstol=1e-5, reltol = 1e-4);
 robertson_plot(sol1, ref_sol, "SSPMPRK22(0.5, 1.0)")
 ```
 
-For comparisons with schemes from [OrdinaryDiffEq.jl](https://docs.sciml.ai/OrdinaryDiffEq/stable/), we choose the second-order schemes `MPRK22(0.5)` and `MPRK22(1.0)` as well as the third-order scheme `MPRK43I(0.5, 0.75)`.
+For comparisons with schemes from [OrdinaryDiffEq.jl](https://docs.sciml.ai/OrdinaryDiffEq/stable/), we choose the second-order schemes `MPRK22(0.5)` and `MPRK22(1.0)`, the third-order scheme `MPRK43I(0.5, 0.75)` and the 10th order scheme `MPDeC(10)`.
 
 ```@example ROBER
[…]
