+++ Project.toml
@@ -5,6 +5,7 @@ version = "0.2.13-DEV"
 
 [deps]
 FastBroadcast = "7034ab61-46d4-4ed7-9d0f-46aef9175898"
+JuMP = "4076af6c-e467-56ae-b986-b466b2749572"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 LinearSolve = "7ed4a6bd-45f5-4d41-b270-4a48e9bafcae"
 MuladdMacro = "46d2c3a1-f734-5fdb-9937-b9b9aeba4221"
@@ -20,6 +21,7 @@ SymbolicIndexingInterface = "2efcf032-c050-4f8e-a9bb-153293bab1f5"
 
 [compat]
 FastBroadcast = "0.3.5"
+JuMP = "1.26.0"
 LinearAlgebra = "1"
 LinearSolve = "3.7.1"
 MuladdMacro = "0.2.4"
