+++ src/PositiveIntegrators.jl
@@ -42,10 +42,6 @@ import OrdinaryDiffEqCore: alg_order, isfsal,
 
 using RecipesBase: @recipe
 
-using JuMP: Model, @variable, @objective, @constraint, print, objective_value, set_silent,
-            optimize!, is_solved_and_feasible, value
-using SciMLBase: DiscreteCallback
-
 # 2. Export functionality defining the public API
 export PDSFunction, PDSProblem
 export ConservativePDSFunction, ConservativePDSProblem
@@ -63,7 +59,9 @@ export work_precision_adaptive, work_precision_adaptive!, work_precision_fixed,
        work_precision_fixed!
 export rel_max_error_overall, rel_max_error_tend, rel_l1_error_tend, rel_l2_error_tend
 
+# implemented in SanduProjectionExt
 export SanduProjection
+abstract type SanduProjection end
 
 # 3. Load source code
 
@@ -86,7 +84,7 @@ include("interpolation.jl")
 include("PDSProblemLibrary.jl")
 
 # Sandu projection
-include("sandu_projection.jl")
+#include("sandu_projection.jl")
 
 # additional auxiliary functions
 include("utilities.jl")
