### 2025-05-23 #8130 @AndresYague
@junghans I am of the same opinion as @svchb . Except for a few trivial nitpicks on the paper wording, the work done on this package meets or exceeds the requirements for a JOSS publication, therefore I also recommend the publication of this submission.
