### 2025-05-23 #8130 @AndresYague
@junghans Sure!

-Summary

1)

"Currently, the package provides MPRK, SSP-MPRK, and MPDeC schemes, all of which are unconditionally positive and also preserve the conservation property when applied to a conservative system."

Could be reworded to highlight the most important point of this package:

"Currently, the package provides **the unconditionally positivity preserving** MPRK, SSP-MPRK, and MPDeC schemes, which are also conservative when applied to conservative systems."

-Body:

2)

"Many systems of ordinary differential equations that model real-life applications have positive solutions, and it is quite natural to require that numerical solutions of such systems also remain positive. For some of these systems unconditionally positivity-preserving time integration methods are helpful or even necessary to obtain meaningful solutions."

I do not think the use of "natural" is warranted here, I would reword this to:

"Many systems of ordinary differential equations that model real-life applications have positive solutions. For some of these systems, unconditionally positivity-preserving time integration methods are helpful or even necessary to obtain meaningful solutions."

If we want to emphasize the importance of these methods, an example (any example, this is just my take) could be added here:

"As an example, in reaction networks, it is non-physical to have negative abundances".

3)

"Unfortunately, positivity is a property that almost all standard time integration schemes, such as Runge–Kutta methods, Rosenbrock methods, or linear multistep methods, do not preserve."

This sentence should be turned around for clarity:

"Unfortunately, positivity is not preserved in almost any of the standard time integration schemes, such as Runge–Kutta methods, Rosenbrock methods, or linear multistep methods."

4)

"However, this is only first-order accurate and, in addition, the preservation of positivity within the nonlinear iteration process poses a problem."

This should be shortened:

"However, this is only first-order accurate and the preservation of positivity within the nonlinear iteration process poses a problem."

An explanation should be added as to why "the nonlinear iteration process poses a problem".

And those are all my nitpicks.
