+++ paper/paper.md
@@ -24,26 +24,28 @@ affiliations:
    index: 2
  - name: Institute of Mathematics, Johannes Gutenberg University Mainz, Germany
    index: 3
-date: 23 April 2025
+date: 24 May 2025
 bibliography: paper.bib
 ---
 
 # Summary
 
-We introduce PositiveIntegrators.jl, a Julia package that provides efficient implementations of various time integration schemes for the solution of positive ordinary differential equations, making these methods accessible for users and comparable for researchers. Currently, the package provides MPRK, SSP-MPRK, and MPDeC schemes, all of which are unconditionally positive and also preserve the conservation property when applied to a conservative system.
+We introduce PositiveIntegrators.jl, a Julia package that provides efficient implementations of various time integration schemes for the solution of positive ordinary differential equations, making these methods accessible for users and comparable for researchers.
+Currently, the package provides the the unconditionally positivity-preserving MPRK, SSP-MPRK, and MPDeC schemes, which are also conservative when applied to a conservative system.
 The package is fully compatible with DifferentialEquations.jl, which allows a direct comparison between the provided schemes and standard methods.
 
 
 # Statement of need
 
-Many systems of ordinary differential equations that model real-life applications have positive solutions, and it is quite natural to require that numerical solutions of such systems also remain positive.
-For some of these systems unconditionally positivity-preserving time integration methods are helpful or even necessary to obtain meaningful solutions. 
+Many systems of ordinary differential equations that model real-life applications have positive solutions.
+For some of these systems unconditionally positivity-preserving time integration methods are helpful or even necessary to obtain meaningful solutions.
 
-Unfortunately, positivity is a property that almost all standard time integration schemes, such as Runge–Kutta methods, Rosenbrock methods, or linear multistep methods, do not preserve.
+Unfortunately, positivity is not preserved in almost any of the standard time integration schemes, such as Runge–Kutta methods, Rosenbrock methods, or linear multistep methods.
 In particular, higher-order general linear methods cannot preserve positivity unconditionally [@bolley1978conservation].
 The only standard scheme with which unconditional positivity can be achieved is the implicit Euler method
-(assuming that the nonlinear systems are solved exactly). However, this is only first-order accurate and, in addition, the preservation of positivity within the nonlinear iteration process poses a problem. 
-Another strategy for preserving positivity used in existing open source or commercial packages (like MATLAB) is to set negative solution components that are accepted by the step size control to zero. Unfortunately, this can have a negative impact on possible conservation properties. Further approaches in the literature include projections in between time steps [@sandu2001positive; @nusslein2021positivity], if a negative solution was computed, or it is tried to reduce the time step size as long as a nonnegative solution is calculated. Finally, strong stability preserving (SSP) methods can also be used to preserve positivity, but this is again subject to step size limitations [@gottlieb2011strong]. 
+(assuming that the nonlinear systems are solved exactly).
+However, this is only first-order accurate and the preservation of positivity within the nonlinear iteration process poses a problem.
[…]
