### 2025-05-24 09e44d4
comment on issues with nonlinear solvers
