+++ paper/paper.md
@@ -44,7 +44,7 @@ Unfortunately, positivity is not preserved in almost any of the standard time in
 In particular, higher-order general linear methods cannot preserve positivity unconditionally [@bolley1978conservation].
 The only standard scheme with which unconditional positivity can be achieved is the implicit Euler method
 (assuming that the nonlinear systems are solved exactly).
-However, this is only first-order accurate and the preservation of positivity within the nonlinear iteration process poses a problem.
+However, this is only first-order accurate and the preservation of positivity within the nonlinear iteration process poses a problem. For example, it may happen that the right-hand side of the differential equation is only defined for nonnegative values and would throw an error if a negative value is passed to it, e.g., in the case of a square root. Standard nonlinear solvers like Newton's method or fixed-point iterations will in general not only produce iterates that are nonnegative, causing the iteration to fail.
 Another strategy for preserving positivity used in existing open source or commercial packages (like MATLAB) is to set negative solution components that are accepted by the step size control to zero. Unfortunately, this can have a negative impact on possible conservation properties. Further approaches in the literature include projections in between time steps [@sandu2001positive; @nusslein2021positivity], if a negative solution was computed, or it is tried to reduce the time step size as long as a nonnegative solution is calculated. Finally, strong stability preserving (SSP) methods can also be used to preserve positivity, but this is again subject to step size limitations [@gottlieb2011strong].
 
 Consequently, various new, unconditionally positive schemes have been introduced in recent years, see @burchard2003, @Bruggeman2007, @Broekhuizen2008, @Formaggia2011, @Ortleb2017, @kopeczmeister2018order2, @kopeczmeister2018order3, @huang2019order2, @huang2019order3, @OeffnerTorlo2020, @Martiradonna2020, @Avila2020, @Avila2021, @Blanes2022, @Zhu2024, @Izzo2025, and @Izgin2025. Among these, most of the literature is devoted to modified Patankar--Runge--Kutta (MPRK) methods.
