### 2025-05-26 36166f6
added linear Invariants for linmod test problem, typo
