+++ src/proddest.jl
@@ -132,8 +132,12 @@ function PDSFunction{iip, FullSpecialize}(P, D;
     end
     PDSFunction{iip, FullSpecialize, typeof(P), typeof(D), typeof(p_prototype),
                 typeof(d_prototype),
-                typeof(std_rhs), typeof(analytic), typeof(lin_invariants)}(P, D, p_prototype, d_prototype, std_rhs,
-                                                   analytic, lin_invariants)
+                typeof(std_rhs), typeof(analytic), typeof(lin_invariants)}(P, D,
+                                                                           p_prototype,
+                                                                           d_prototype,
+                                                                           std_rhs,
+                                                                           analytic,
+                                                                           lin_invariants)
 end
 
 # Evaluation of a PDSFunction
@@ -318,8 +322,11 @@ function ConservativePDSFunction{iip, FullSpecialize}(P;
         std_rhs = ConservativePDSStdRHS(P, p_prototype)
     end
     ConservativePDSFunction{iip, FullSpecialize, typeof(P), typeof(p_prototype),
-                            typeof(std_rhs), typeof(analytic), typeof(lin_invariants)}(P, p_prototype, std_rhs,
-                                                               analytic, lin_invariants)
+                            typeof(std_rhs), typeof(analytic), typeof(lin_invariants)}(P,
+                                                                                       p_prototype,
+                                                                                       std_rhs,
+                                                                                       analytic,
+                                                                                       lin_invariants)
 end
 
 # Evaluation of a ConservativePDSFunction
