### 2025-05-27 485e354
add lin invariants to problem library
