+++ src/PDSProblemLibrary.jl
@@ -33,7 +33,7 @@ There is one independent linear invariant, e.g. ``u_1+u_2 = 1``.
 """
 prob_pds_linmod = ConservativePDSProblem(P_linmod, u0_linmod, (0.0, 2.0),
                                          analytic = f_linmod_analytic, std_rhs = f_linmod,
-                                         lin_invariants = [1.0 1.0])
+                                         lin_invariants = @SMatrix[1.0 1.0])
 
 function P_linmod!(P, u, p, t)
     P .= P_linmod(u, p, t)
@@ -52,7 +52,8 @@ Same as [`prob_pds_linmod`](@ref) but with in-place computation.
 prob_pds_linmod_inplace = ConservativePDSProblem(P_linmod!, Array(u0_linmod),
                                                  (0.0, 2.0),
                                                  analytic = f_linmod_analytic,
-                                                 std_rhs = f_linmod!)
+                                                 std_rhs = f_linmod!,
+                                                 lin_invariants = @SMatrix[1.0 1.0])
 
 # nonlinear model problem
 function P_nonlinmod(u, p, t)
@@ -87,7 +88,8 @@ There is one independent linear invariant, e.g. ``u_1+u_2+u_3 = 10.0``.
   [DOI: 10.1016/S0168-9274(03)00101-6](https://doi.org/10.1016/S0168-9274(03)00101-6)
 """
 prob_pds_nonlinmod = ConservativePDSProblem(P_nonlinmod, u0_nonlinmod, (0.0, 30.0),
-                                            std_rhs = f_nonlinmod)
+                                            std_rhs = f_nonlinmod,
+                                            lin_invariants = @SMatrix[1.0 1.0 1.0])
 
 # robertson problem
 function P_robertson(u, p, t)
@@ -120,7 +122,8 @@ There is one independent linear invariant, e.g. ``u_1+u_2+u_3 = 1.0``.
   2nd Edition, Springer (2002): Section IV.1.
 """
 prob_pds_robertson = ConservativePDSProblem(P_robertson, u0_robertson, (0.0, 1.0e11),
-                                            std_rhs = f_robertson)
+                                            std_rhs = f_robertson,
+                                            lin_invariants = @SMatrix[1.0 1.0 1.0])
 
 # brusselator problem
 function P_brusselator(u, p, t)
@@ -167,7 +170,9 @@ There are two independent linear invariants, e.g. ``u_1+u_4+u_5+u_6 = 10.2`` and
   [DOI: 10.1007/s10915-016-0267-9](https://doi.org/10.1007/s10915-016-0267-9)
 """
 prob_pds_brusselator = ConservativePDSProblem(P_brusselator, u0_brusselator, (0.0, 10.0),
-                                              std_rhs = f_brusselator)
+                                              std_rhs = f_brusselator,
+                                              lin_invariants = @SMatrix[1.0 0.0 0.0 1.0 1.0 1.0;
+                                                                        0.0 1.0 1.0 0.0 0.0 0.0])
 
 # SIR problem
 P_sir(u, p, t) = @SMatrix [0.0 0.0 0.0; 2*u[1]*u[2] 0.0 0.0; 0.0 u[2] 0.0]
@@ -196,7 +201,8 @@ There is one independent linear invariant, e.g. ``u_1+u_2+u_3 = 1.0``.
[…]
