### 2025-05-27 925588c
renamed lin_invariants to linear_invariants, corrected minmapk linear_invariants matrix
