+++ src/proddest.jl
@@ -51,7 +51,7 @@ struct PDSFunction{iip, specialize, P, D, PrototypeP, PrototypeD, StdRHS, Ta, LI
     d_prototype::PrototypeD
     std_rhs::StdRHS
     analytic::Ta
-    lin_invariants::LI
+    linear_invariants::LI
 end
 
 # define behavior of PDSFunctions for non-existing fields
@@ -92,7 +92,7 @@ function PDSProblem{iip}(P, D, u0, tspan, p = NullParameters();
                          p_prototype = nothing,
                          analytic = nothing,
                          std_rhs = nothing,
-                         lin_invariants = nothing,
+                         linear_invariants = nothing,
                          kwargs...) where {iip}
 
     # p_prototype is used to store evaluations of P, if P is in-place.
@@ -104,7 +104,7 @@ function PDSProblem{iip}(P, D, u0, tspan, p = NullParameters();
     d_prototype = similar(u0 ./ oneunit(first(tspan)))
 
     PD = PDSFunction{iip}(P, D; p_prototype, d_prototype,
-                          analytic, std_rhs, lin_invariants)
+                          analytic, std_rhs, linear_invariants)
     PDSProblem{iip}(PD, u0, tspan, p; kwargs...)
 end
 
@@ -126,18 +126,18 @@ function PDSFunction{iip, FullSpecialize}(P, D;
                                           d_prototype = nothing,
                                           analytic = nothing,
                                           std_rhs = nothing,
-                                          lin_invariants = nothing) where {iip}
+                                          linear_invariants = nothing) where {iip}
     if std_rhs === nothing
         std_rhs = PDSStdRHS(P, D, p_prototype, d_prototype)
     end
     PDSFunction{iip, FullSpecialize, typeof(P), typeof(D), typeof(p_prototype),
                 typeof(d_prototype),
-                typeof(std_rhs), typeof(analytic), typeof(lin_invariants)}(P, D,
-                                                                           p_prototype,
-                                                                           d_prototype,
-                                                                           std_rhs,
-                                                                           analytic,
-                                                                           lin_invariants)
+                typeof(std_rhs), typeof(analytic), typeof(linear_invariants)}(P, D,
+                                                                              p_prototype,
+                                                                              d_prototype,
+                                                                              std_rhs,
+                                                                              analytic,
+                                                                              linear_invariants)
 end
 
 # Evaluation of a PDSFunction
@@ -253,7 +253,7 @@ struct ConservativePDSFunction{iip, specialize, P, PrototypeP, StdRHS, Ta, LI} <
     p_prototype::PrototypeP # prototype for production terms
     std_rhs::StdRHS # standard right-hand side evaluation function
     analytic::Ta # analytic solution (or nothing)
-    lin_invariants::LI
+    linear_invariants::LI
 end
 
 # define behavior of ConservativePDSFunction for non-existing fields
@@ -288,7 +288,7 @@ function ConservativePDSProblem{iip}(P, u0, tspan, p = NullParameters();
                                      p_prototype = nothing,
                                      analytic = nothing,
                                      std_rhs = nothing,
-                                     lin_invariants = nothing,
+                                     linear_invariants = nothing,
                                      kwargs...) where {iip}
 
     # p_prototype is used to store evaluations of P, if P is in-place.
@@ -296,7 +296,7 @@ function ConservativePDSProblem{iip}(P, u0, tspan, p = NullParameters();
         p_prototype = zeros(eltype(u0), (length(u0), length(u0))) / oneunit(first(tspan))
     end
 
[…]
