### 2025-05-28 253b7a8
add tests for linear invariant matrices
