### 2025-05-28 e2a499e
add @SMatrix, typo
