### 2025-05-29 b8bedff
described 'linear invariants' in docstring
