### 2025-05-29 #8130 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
