### 2025-05-30 fdb86a5
revised docstring
