### 2025-05-30 1e80944
Store linear invariants matrix in (Conservative)PDSProblem (#163)
* added lin invariants in PDSFunction and ConservativePDSFunction

* added linear Invariants for linmod test problem, typo

* add lin invariants to problem library

* renamed lin_invariants to linear_invariants, corrected minmapk linear_invariants matrix

* add tests for linear invariant matrices

* typo

* add @SMatrix, typo

* format

* fix stefan

* requested changes hendrik 'runtests'

* described 'linear invariants' in docstring

* revised docstring
