+++ src/proddest.jl
@@ -5,7 +5,8 @@ abstract type AbstractPDSProblem end
     PDSProblem(P, D, u0, tspan, p = NullParameters();
                p_prototype = nothing,
                analytic = nothing,
-               std_rhs = nothing)
+               std_rhs = nothing,
+               linear_invariants = nothing)
 
 A structure describing a system of ordinary differential equations in form of a production-destruction system (PDS).
 `P` denotes the function defining the production matrix ``P``.
@@ -31,6 +32,9 @@ The functions `P` and `D` can be used either in the out-of-place form with signa
   the production-destruction representation of the ODE, will use this function
   instead to compute the solution. If not specified,
   a default implementation calling `P` and `D` is used.
+-`linear_invariants`: The rows of this matrix contain the linear invariants of the ODE. 
+  Certain solvers or callbacks require this matrix.
+  Note that this feature is experimental and its API may change in future releases.
 
 ## References
 
@@ -43,7 +47,7 @@ The functions `P` and `D` can be used either in the out-of-place form with signa
 struct PDSProblem{iip} <: AbstractPDSProblem end
 
 # New ODE function PDSFunction
-struct PDSFunction{iip, specialize, P, D, PrototypeP, PrototypeD, StdRHS, Ta} <:
+struct PDSFunction{iip, specialize, P, D, PrototypeP, PrototypeD, StdRHS, Ta, LI} <:
        AbstractODEFunction{iip}
     p::P
     d::D
@@ -51,6 +55,7 @@ struct PDSFunction{iip, specialize, P, D, PrototypeP, PrototypeD, StdRHS, Ta} <:
     d_prototype::PrototypeD
     std_rhs::StdRHS
     analytic::Ta
+    linear_invariants::LI
 end
 
 # define behavior of PDSFunctions for non-existing fields
@@ -73,8 +78,7 @@ function Base.getproperty(obj::PDSFunction, sym::Symbol)
 end
 
 # Most general constructor for PDSProblems
-function PDSProblem(P, D, u0, tspan, p = NullParameters();
-                    kwargs...)
+function PDSProblem(P, D, u0, tspan, p = NullParameters(); kwargs...)
     Piip = isinplace(P, 4)
     Diip = isinplace(D, 4)
     if Piip == Diip
@@ -87,11 +91,16 @@ end
 
 # Specialized constructor for PDSProblems setting `iip` manually
 # (arbitrary functions)
-function PDSProblem{iip}(P, D, u0, tspan, p = NullParameters();
+function PDSProblem{iip}(P,
+                         D,
+                         u0,
+                         tspan,
+                         p = NullParameters();
                          p_prototype = nothing,
                          analytic = nothing,
                          std_rhs = nothing,
-                         kwargs...) where {iip}
+                         linear_invariants = nothing,
+                         kwargs...,) where {iip}
 
     # p_prototype is used to store evaluations of P, if P is in-place.
     if isnothing(p_prototype) && iip
@@ -101,8 +110,8 @@ function PDSProblem{iip}(P, D, u0, tspan, p = NullParameters();
     # evaluations of D.
     d_prototype = similar(u0 ./ oneunit(first(tspan)))
 
-    PD = PDSFunction{iip}(P, D; p_prototype, d_prototype,
-                          analytic, std_rhs)
+    PD = PDSFunction{iip}(P, D; p_prototype, d_prototype, analytic, std_rhs,
+                          linear_invariants)
     PDSProblem{iip}(PD, u0, tspan, p; kwargs...)
 end
 
@@ -119,18 +128,26 @@ function PDSFunction{iip}(P, D; kwargs...) where {iip}
 end
 
 # Most specific constructor for PDSFunction
-function PDSFunction{iip, FullSpecialize}(P, D;
[…]
